requirements.txt
numpy
tensorflow
Pillow
scikit-image
tqdm
openslide-python
openslide-bin
h5py
opencv-python-headless
clearml
boto3
